refactor: getChart server action으로 수정
